Cleanups/Refactorings
