+++ Core/World/Room.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using ECS2022_23.Core.Entities;
+using ECS2022_23.Core.Loader;
 using ECS2022_23.Enums;
 using ECS2022_23.Helper;
 using Microsoft.Xna.Framework;
