+++ Core/World/StageGenerator.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
 using System.Linq;
+using ECS2022_23.Core.Loader;
 using ECS2022_23.Enums;
 using Microsoft.Xna.Framework;
 using TiledCS;
