+++ ECS2022_23.csproj
@@ -11,9 +11,9 @@
         <ApplicationIcon>Icon.ico</ApplicationIcon>
     </PropertyGroup>
     <ItemGroup>
-        <None Remove="Icon.ico"/>
-        <None Remove="Icon.bmp"/>
-        <None Include="Assets\World\tilesets\images\tileset001_image_purple.png"/>
+        <None Remove="Icon.ico" />
+        <None Remove="Icon.bmp" />
+        <None Include="Assets\World\tilesets\images\tileset001_image_purple.png" />
     </ItemGroup>
     <ItemGroup>
         <EmbeddedResource Include="Icon.ico">
@@ -24,10 +24,10 @@
         </EmbeddedResource>
     </ItemGroup>
     <ItemGroup>
-        <PackageReference Include="Comora" Version="0.5.0"/>
-        <PackageReference Include="MonoGame.Framework.DesktopGL" Version="3.8.1.303"/>
-        <PackageReference Include="MonoGame.Content.Builder.Task" Version="3.8.1.303"/>
-        <PackageReference Include="TiledCS" Version="3.3.3"/>
+        <PackageReference Include="Comora" Version="0.5.0" />
+        <PackageReference Include="MonoGame.Framework.DesktopGL" Version="3.8.1.303" />
+        <PackageReference Include="MonoGame.Content.Builder.Task" Version="3.8.1.303" />
+        <PackageReference Include="TiledCS" Version="3.3.3" />
     </ItemGroup>
     <ItemGroup>
         <Reference Include="TiledPipelineReader, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null">
@@ -35,17 +35,17 @@
         </Reference>
     </ItemGroup>
     <ItemGroup>
-        <Content Include="Assets\GameStateManagement\blank.png"/>
-        <Content Include="Assets\GameStateManagement\controls.png"/>
-        <Content Include="Assets\Sprites\astro.png"/>
-        <Content Include="Assets\Sprites\spritesheet.png"/>
-        <Content Include="Assets\Sprites\spritesheet_red.png"/>
+        <Content Include="Assets\GameStateManagement\blank.png" />
+        <Content Include="Assets\GameStateManagement\controls.png" />
+        <Content Include="Assets\Sprites\astro.png" />
+        <Content Include="Assets\Sprites\spritesheet.png" />
+        <Content Include="Assets\Sprites\spritesheet_red.png" />
     </ItemGroup>
     <ItemGroup>
-        <Folder Include="Assets\GameStateManagement\Sounds"/>
+        <Folder Include="Assets\GameStateManagement\Sounds" />
     </ItemGroup>
     <Target Name="RestoreDotnetTools" BeforeTargets="Restore">
-        <Message Text="Restoring dotnet tools" Importance="High"/>
-        <Exec Command="dotnet tool restore"/>
+        <Message Text="Restoring dotnet tools" Importance="High" />
+        <Exec Command="dotnet tool restore" />
     </Target>
 </Project>
